refactor: rename core store file
