+++ examples/iceberg/tsconfig.json
@@ -21,7 +21,7 @@
 		],
 		"paths": {
 			"@polar/polar": ["../../src/core/index.ts"],
-			"@polar/polar/store": ["../../src/core/stores/export.ts"],
+			"@polar/polar/store": ["../../src/core/stores/index.ts"],
 			"@polar/polar/plugins/*": ["../../src/plugins/*/index.ts"],
 			"@polar/polar/plugins/*/store": ["../../src/plugins/*/store.ts"],
 			"@/*": ["../../src/*"]
