+++ src/components/PolarIconButton.ce.vue
@@ -23,7 +23,7 @@
 <script setup lang="ts">
 import { storeToRefs } from 'pinia'
 
-import { useCoreStore } from '@/core/stores/export'
+import { useCoreStore } from '@/core/stores'
 
 defineProps<{
 	hint: string
