+++ src/core/utils/export/store.ts
@@ -1,7 +1,7 @@
 import { watch, type WatchOptions } from 'vue'
 
 import type { PolarContainer } from '@/core'
-import type { useCoreStore } from '@/core/stores/export'
+import type { useCoreStore } from '@/core/stores'
 import type {
 	BundledPluginId,
 	BundledPluginStores,
