+++ src/lib/notifyUser.ts
@@ -1,6 +1,6 @@
 import type { Ref } from 'vue'
 
-import { useCoreStore } from '@/core/stores/export'
+import { useCoreStore } from '@/core/stores'
 import { computedT } from '@/lib/computedT'
 import { type ToastOptions } from '@/plugins/toast'
