+++ src/plugins/geoLocation/components/GeoLocation.ce.vue
@@ -16,7 +16,7 @@ import { storeToRefs } from 'pinia'
 import { computed } from 'vue'
 
 import PolarIconButton from '@/components/PolarIconButton.ce.vue'
-import { useCoreStore } from '@/core/stores/export'
+import { useCoreStore } from '@/core/stores'
 
 import { useGeoLocationStore } from '../store'
 import { PluginId } from '../types'
