+++ src/plugins/iconMenu/components/StandardMenuList.ce.vue
@@ -47,7 +47,7 @@ import {
 } from 'vue'
 
 import PolarIconButton from '@/components/PolarIconButton.ce.vue'
-import { useCoreStore } from '@/core/stores/export'
+import { useCoreStore } from '@/core/stores'
 
 import type { Menu } from '../types'
