+++ src/plugins/loadingIndicator/store.ts
@@ -7,7 +7,7 @@
 import { defineStore } from 'pinia'
 import { computed, ref } from 'vue'
 
-import { useCoreStore } from '@/core/stores/export.ts'
+import { useCoreStore } from '@/core/stores'
 
 const styles = [
 	'KernLoader',
