+++ src/plugins/toast/components/ToastContainer.ce.vue
@@ -8,7 +8,7 @@
 <script setup lang="ts">
 import { storeToRefs } from 'pinia'
 
-import { useCoreStore } from '@/core/stores/export'
+import { useCoreStore } from '@/core/stores'
 
 import ToastUI from './ToastUI.ce.vue'
