+++ src/plugins/toast/store.ts
@@ -8,7 +8,7 @@ import { toMerged } from 'es-toolkit'
 import { acceptHMRUpdate, defineStore } from 'pinia'
 import { computed, ref, toRaw, type Reactive } from 'vue'
 
-import { useCoreStore } from '@/core/stores/export'
+import { useCoreStore } from '@/core/stores'
 
 import {
 	PluginId,
@@ -113,7 +113,7 @@ if (import.meta.vitest) {
 	const { expect, test: _test, vi } = import.meta.vitest
 	const { createPinia, setActivePinia } = await import('pinia')
 	const { reactive } = await import('vue')
-	const useCoreStoreFile = await import('@/core/stores/export')
+	const useCoreStoreFile = await import('@/core/stores')
 
 	/* eslint-disable no-empty-pattern */
 	const test = _test.extend<{
