+++ typedoc.json
@@ -3,7 +3,7 @@
 	"entryPoints": [
 		"src/core/index.ts",
 		"src/core/locales.ts",
-		"src/core/stores/export.ts",
+		"src/core/stores/index.ts",
 		"src/plugins/*/index.ts",
 		"src/plugins/*/locales.ts",
 		"src/plugins/*/store.ts"
